requirements.txt
streamlit==1.31.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
requests==2.31.0
python-dotenv==1.0.1
SQLAlchemy==2.0.25
pandas==2.2.0
openai==1.12.0
PyPDF2==3.0.1
